Rebase into previous commit later once finished debugging.
Unfortunately it does not look like I can save the state of the dataset within an epoch using the dataset iterator because the random sampling mapping has 'state' that cannot be captured in the checkpoint. We can still save checkpoints at each iteration however. And then if we crash within an epoch, I added a new variable such that we start at the same iteration. Since we set the seed of shuffle, this should mean we start at the same place in the dataset, but I am not 100% sure about that.